requirements.txt
astroid==2.8.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
firebase-admin==5.0.3
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.2.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.3
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
httplib2==0.19.1
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.2
packaging==21.0
platformdirs==2.4.0
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.11.1
pymongo==3.12.0
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
wrapt==1.12.1
